+++ packages/furo-data/lib/FieldNode.js
@@ -65,8 +65,8 @@ export class FieldNode extends EventTreeNode {
         if(val.hasOwnProperty(field._name)){
           field.value = val[field._name];
         }
-
       }
+        this.dispatchNodeEvent(new NodeEvent('branch-value-changed', this, false));
     } else {
       this.oldvalue = this.value;
       this._value = val;
@@ -102,12 +102,6 @@ export class FieldNode extends EventTreeNode {
     }
     return this._value;
   }
-
-  set(val) {
-    this.oldvalue = this.value;
-    this.value = val;
-
-  };
 
   _clearInvalidity() {
     if (!this._isValid) {