+++ packages/furo-experiments/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.8](https://github.com/veith/FuroBaseComponents/compare/@furo/experiments@0.0.7...@furo/experiments@0.0.8) (2019-07-11)
+
+**Note:** Version bump only for package @furo/experiments
+
+
+
+
+
+## [0.0.7](https://github.com/veith/FuroBaseComponents/compare/@furo/experiments@0.0.6...@furo/experiments@0.0.7) (2019-07-09)
+
+**Note:** Version bump only for package @furo/experiments
+
+
+
+
+
+## [0.0.6](https://github.com/veith/FuroBaseComponents/compare/@furo/experiments@0.0.5...@furo/experiments@0.0.6) (2019-07-03)
+
+**Note:** Version bump only for package @furo/experiments
+
+
+
+
+
 ## [0.0.5](https://github.com/veith/FuroBaseComponents/compare/@furo/experiments@0.0.4...@furo/experiments@0.0.5) (2019-07-01)
 
 **Note:** Version bump only for package @furo/experiments